Fix formatting and placement
